[product,like] 상품, 좋아요 entity , dto 수정

+++ src/main/java/com/example/nowpt/mvc/dto/ProductLikeDto.java
@@ -9,7 +9,7 @@ import lombok.NonNull;
 @NoArgsConstructor
 public class ProductLikeDto {
     @NonNull
-    private int productSn;
+    private long productSn;
 
     @NonNull
     private String memberEmail;
